# docs/swagger.yaml
basePath: /
definitions:
  http.Response:
    properties:
      data: {}
      error_code:
        type: string
      successful:
        type: boolean
    type: object
  port.CreateProductRequest:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
      sale_price:
        type: number
    required:
    - name
    - price
    type: object
  port.PatchProductRequest:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
      sale_price:
        type: number
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for managing products.
  title: Go Pre-Test API
  version: "1.0"
paths:
  /product:
    post:
      consumes:
      - application/json
      description: Create a new product with the input payload
      parameters:
      - description: Product to create
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/port.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Response'
      summary: Create a new product
      tags:
      - product
  /product/{id}:
    patch:
      consumes:
      - application/json
      description: Update specific fields of a product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/port.PatchProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Response'
      summary: Patch a product
      tags:
      - product
swagger: "2.0"